- Set up to scan TestStuie folder which is now installed in test area. - Set source URI on library import and skip when scanning for validation - Fix file name search a bit.
