+++ source/MaterialXTest/CMakeLists.txt
@@ -32,6 +32,9 @@ if (MATERIALX_BUILD_VIEW)
   add_custom_command(TARGET MaterialXTest POST_BUILD
                    COMMAND ${CMAKE_COMMAND} -E copy_directory
                    ${CMAKE_SOURCE_DIR}/documents/Geometry ${CMAKE_CURRENT_BINARY_DIR}/documents/Geometry)
+  add_custom_command(TARGET MaterialXTest POST_BUILD
+                   COMMAND ${CMAKE_COMMAND} -E copy_directory
+                   ${CMAKE_SOURCE_DIR}/documents/TestSuite ${CMAKE_CURRENT_BINARY_DIR}/documents/TestSuite)
 endif(MATERIALX_BUILD_VIEW)
                    
 set_target_properties(
